Delete fixed
Fixed the delete button.
It only works if current num is different from 0.
after deleting a number it checks if the currentNum is empty or not.
If current num is empty It checks if there's an operator. If there isn't an operator than Display inner text and current number both are set to 0.
Otherwise, if there is an operator, the program deleltes the operator, sets the currentNum iquals to the displayed value and also sets the firstNum to an empty string